[CHORE] LetterState 위치 변경(#13)
